Finish extracting models into the domain subdirectory.
