+++ world/dominion/admin.py
@@ -6,9 +6,9 @@ from django.db.models import Q
 from django.shortcuts import reverse
 from django.utils.html import escape
 
-from .models import (PlayerOrNpc, Organization, Domain, Agent, AgentOb, Minister, MapLocation,
-                     AssetOwner, Region, Land, Castle, WorkSetting, PraiseOrCondemn,
-                     Ruler, Member, Task, CraftingRecipe, CraftingMaterialType, CraftingMaterials,
+from .models import (PlayerOrNpc, Organization, Agent, AgentOb, MapLocation,
+                     AssetOwner, Region, Land,  WorkSetting, PraiseOrCondemn,
+                     Member, Task, CraftingRecipe, CraftingMaterialType, CraftingMaterials,
                      RPEvent, AccountTransaction, AssignedTask,
                      OrgRelationship, Reputation, TaskSupporter, InfluenceCategory,
                      Renown, SphereOfInfluence, TaskRequirement, ClueForOrg,
@@ -18,7 +18,7 @@ from .models import (PlayerOrNpc, Organization, Domain, Agent, AgentOb, Minister
 from world.dominion.plots.models import (Plot, PlotAction, PlotUpdate, ActionOOCQuestion,
                                          PCPlotInvolvement, OrgPlotInvolvement, PlotActionAssistant)
 
-from world.dominion.domain.models import (Army, Orders, MilitaryUnit, OrgUnitModifiers)
+from world.dominion.domain.models import (Domain, Castle, Minister, Army, Ruler, Orders, MilitaryUnit, OrgUnitModifiers)
 
 from web.help_topics.templatetags.app_filters import mush_to_html
 from world.exploration.models import Shardhaven, ShardhavenType
