+++ world/dominion/domain/unit_types.py
@@ -5,9 +5,9 @@ All the stats for different kinds of military units are defined here and
 will be used at runtime.
 """
 import traceback
-from .combat_grid import PositionActor
+from world.dominion.combat_grid import PositionActor
 from random import randint
-from . import unit_constants
+from world.dominion.domain import unit_constants
 
 _UNIT_TYPES = {}
