+++ world/dominion/dominion_constants.py
@@ -1,0 +1,26 @@
+BASE_WORKER_COST = 0.10
+SILVER_PER_BUILDING = 225.00
+FOOD_PER_FARM = 100.00
+# default value for a global modifier to Dominion income, can be set as a ServerConfig value on a per-game basis
+DEFAULT_GLOBAL_INCOME_MOD = -0.25
+# each point in a dominion skill is a 5% bonus
+BONUS_PER_SKILL_POINT = 0.10
+# number of workers for a building to be at full production
+SERFS_PER_BUILDING = 20.0
+# population cap for housing
+POP_PER_HOUSING = 1000
+BASE_POP_GROWTH = 0.01
+DEATHS_PER_LAWLESS = 0.0025
+LAND_SIZE = 10000
+LAND_COORDS = 9
+LIFESTYLES = {
+    0: (-100, -1000),
+    1: (0, 0),
+    2: (100, 2000),
+    3: (200, 3000),
+    4: (500, 4000),
+    5: (1500, 7000),
+    6: (5000, 10000),
+    }
+PRESTIGE_DECAY_AMOUNT = 0.35
+MAX_PRESTIGE_HISTORY = 10
