+++ world/dominion/views.py
@@ -2,8 +2,9 @@
 Views related to the Dominion app
 """
 from django.views.generic import ListView, DetailView, CreateView
-from .models import RPEvent, AssignedTask, Land, Domain, Organization
+from .models import RPEvent, AssignedTask, Land, Organization
 from world.dominion.plots.models import Plot
+from world.dominion.domain.models import Domain
 from .forms import RPEventCommentForm, RPEventCreateForm
 from .view_utils import EventHTMLCalendar
 from django.http import HttpResponseRedirect, HttpResponse
